add test id and assertion ids to payload sent to event bridge
