+++ lib/db/queries/index.js
@@ -1,5 +1,8 @@
 const getAllTests = require('./getAllTests');
 const getTestRuns = require('./getTestRuns');
+const addNewTest = require('./addNewTest');
+const addTestAssertions = require('./addTestAssertions');
+const addTestAlerts = require('./addTestAlerts');
 
 // enables imports elsewhere under a common namespace
 //
@@ -10,4 +13,7 @@ const getTestRuns = require('./getTestRuns');
 module.exports = {
   getAllTests,
   getTestRuns,
+  addNewTest,
+  addTestAssertions,
+  addTestAlerts,
 };
