+++ lib/db/query.js
@@ -1,131 +1,6 @@
 const { dbQuery } = require('./conn');
 const helpers = require('../../utils/helpers');
 
-// TODO: Move to db lookup
-const comparisonIds = {
-  equalTo: 1,
-  notEqualTo: 2,
-  greaterThan: 3,
-  lessThan: 4,
-  greaterThanOrEqualTo: 5,
-  lessThanOrEqualTo: 6,
-  hasKey: 7,
-  notHasKey: 8,
-  hasValue: 9,
-  notHasValue: 10,
-  isEmpty: 11,
-  isNotEmpty: 12,
-  contains: 13,
-  notContains: 14,
-  isNull: 15,
-  isNotNull: 16,
-};
-
-// TODO: Move to db lookup
-const httpMethodIds = {
-  get: 1,
-  post: 2,
-  put: 3,
-  delete: 4,
-  patch: 5,
-  head: 6,
-};
-
-async function addNewTestAlert(testId, alertChannel) {
-  const query1 = `
-    INSERT INTO notification_settings (alerts_on_recovery, alerts_on_failure)
-    VALUES ($1, $2)
-    RETURNING id
-  `;
-
-  const query2 = `
-    INSERT INTO alerts (type, destination, notification_settings_id)
-    VALUES ($1, $2, $3)
-    RETURNING id
-  `;
-
-  const query3 = `
-    INSERT INTO tests_alerts (test_id, alerts_id)
-    VALUES ($1, $2)
-  `;
-
-  let notificationSettingsId;
-  let alertsId;
-  let addAlertResult = false;
-
-  try {
-    const result1 = await dbQuery(
-      query1,
-      alertChannel.alertsOnRecovery,
-      alertChannel.alertsOnFailure,
-    );
-    notificationSettingsId = result1.rows[0].id;
-
-    const result2 = await dbQuery(
-      query2,
-      alertChannel.type,
-      alertChannel.destination,
-      notificationSettingsId,
-    );
-    alertsId = result2.rows[0].id;
-
-    const result3 = await dbQuery(query3, testId, alertsId);
-    addAlertResult = result3.rowCount === 1;
-  } catch (e) {
-    console.error(e);
-  }
-  return addAlertResult;
-}
-
-async function addNewTestAlerts(testId, alertChannels) {
-  if (alertChannels) {
-    try {
-      // eslint-disable-next-line max-len
-      const result = await Promise.all(alertChannels.map((channel) => addNewTestAlert(testId, channel)));
-      console.log('Add New Test Alerts Result: ', result);
-    } catch (e) {
-      console.error(e);
-    }
-  }
-}
-
-async function addNewTestAssertions(testId, assertions) {
-  let query = `
-    INSERT INTO assertions (test_id, type, property, comparison_type_id, expected_value)
-    VALUES
-  `;
-  assertions.forEach((assertion) => {
-    const comparisonId = comparisonIds[assertion.comparison];
-    query += `(${testId}, '${assertion.type}', '${assertion.property || null}', ${comparisonId}, '${assertion.target || null}'),`;
-  });
-
-  let queryCopy = query.slice(0, query.length - 1).replace(/'null'/g, null);
-  queryCopy += ';';
-  const result = await dbQuery(queryCopy);
-  if (result.rowCount === 0) return false;
-  return true;
-}
-
-async function addNewTest(ruleName, RuleArn, test) {
-  const { minutesBetweenRuns, httpRequest, alertChannels } = test;
-  const query = `
-    INSERT INTO tests (name, run_frequency_mins, method_id, url, status, eb_rule_arn) 
-    VALUES ('${ruleName}', '${minutesBetweenRuns}', '${httpMethodIds[httpRequest.method]}', '${httpRequest.url}', 'enabled', '${RuleArn}')
-    RETURNING id
-  `;
-
-  const result = await dbQuery(
-    query,
-  );
-
-  if (result.rowCount === 1) {
-    const { id: testId } = result.rows[0];
-    addNewTestAssertions(testId, httpRequest.assertions);
-    addNewTestAlerts(testId, alertChannels);
-  }
-  return false;
-}
-
 async function editTest(ruleName, RuleArn, test) {
   const updateTestsQuery = `
     UPDATE tests 
@@ -316,7 +191,6 @@ async function deleteTest(testId) {
   console.log('deleted data', data);
 }
 
-module.exports.addNewTest = addNewTest;
 module.exports.getTests = getTests;
 module.exports.getTest = getTest;
 module.exports.getSideload = getSideload;
